removed duplicate code on template list cli
