+++ container_service_extension/client/legacy_native_cluster_api.py
@@ -12,16 +12,6 @@ class LegacyNativeClusterApi:
         self.client = client
         self._uri = f"{self.client.get_api_uri()}/{shared_constants.CSE_URL_FRAGMENT}"  # noqa: E501
 
-    def get_templates(self):
-        method = shared_constants.RequestMethod.GET
-        uri = f"{self._uri}/templates"
-        response = self.client._do_request_prim(
-            method,
-            uri,
-            self.client._session,
-            accept_type='application/json')
-        return process_response(response)
-
     def list_clusters(self, vdc=None, org=None):
         method = shared_constants.RequestMethod.GET
         uri = f"{self._uri}/clusters"
